requirements.txt
aiogram>=3.4
google-generativeai
rapidfuzz
python-dotenv
aiofiles
loguru
Unidecode
aiosqlite
